# cells/erp/api_contract.yaml
...
openapi: 3.0.3
info:
  title: ERP Cell API
  description: 企业资源计划 - 财务总账、应收应付、物料管理、生产计划
  version: 1.0.0
servers: [{ url: /api/v1/erp }]
...
paths:
...
  /orders:
    get:
      tags: [orders]
      parameters: [{ name: X-Tenant-Id, in: header, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
    post:
      tags: [orders]
      parameters: [{ name: X-Request-ID, in: header, required: true }, { name: X-Tenant-Id, in: header, required: true }]
      requestBody: { content: { application/json: { schema: { type: object } } } }
      responses: { '201': { description: Created } }
...
components:
...
  securitySchemes: { BearerAuth: { type: http, scheme: bearer, bearerFormat: JWT } }
security: [{ BearerAuth: [] }]